==> W02_regex_aoi_scraping/step3-merge_presidential_to_one_csv.ipynb <==
import pandas as pd
from selenium import webdriver
from selenium.webdriver.firefox.service import Service
from webdriver_manager.firefox import GeckoDriverManager
import re
import dateparser 

Merge speeches dataframe with presidents dataframe - since there are some slight variations in the president names between these two tables, we will use a fuzzy merge technique

speeches = pd.read_csv("./data/presidential_speeches.csv")
parties = pd.read_csv("./data/presidential_party_and_term.csv")

# change the column name of the presidents dataframe to match the speeches dataframe
# In the presidents dataframe, the column name is 'president'. In the speeches dataframe, the column name is 'name'
parties = parties.rename(columns={'president':'name'})

speeches.head()

parties.head()

import difflib # this will provide us with a 'fuzzy' match between the presidential names found in each table

#convert name in party to name it most closely matches in speeches
parties['name'] = parties['name'].apply(lambda x: difflib.get_close_matches(x, speeches['name'])[0])

# merge the DataFrames into one
merged = speeches.merge(parties)

# view final DataFrame
merged

merged.to_csv("./data/presidential_speeches_merged.csv",index=False)

merged['political party'].value_counts()
